+++ io-type/core/src/commonMain/kotlin/com/wsr/collection/MaxIndexExt.kt
@@ -5,7 +5,7 @@ import com.wsr.IOType
 fun IOType.D1.maxIndex(): Int {
     if (value.isEmpty()) throw IllegalStateException("IOType.D1 is empty")
     var index = 0
-    var max = Double.MIN_VALUE
+    var max = Float.MIN_VALUE
     for (i in value.indices) {
         if (max < this[i]) {
             index = i
