+++ io-type/core/src/commonTest/kotlin/com/wsr/collection/MaxExtTest.kt
@@ -36,7 +36,7 @@ class MaxExtTest {
     fun `D2のmax=全要素の最大値`() {
         // [[1, 2, 3],
         //  [4, 5, 6]]
-        val a = IOType.d2(2, 3) { x, y -> (x * 3 + y + 1).toDouble() }
+        val a = IOType.d2(2, 3) { x, y -> (x * 3 + y + 1).toFloat() }
         val result = a.max()
         assertEquals(
             expected = 6.0,
@@ -48,8 +48,8 @@ class MaxExtTest {
     fun `List_D2のmax=各D2の最大値のリスト`() {
         val list =
             listOf(
-                IOType.d2(2, 2) { x, y -> (x * 2 + y + 1).toDouble() },
-                IOType.d2(2, 2) { x, y -> (x * 2 + y + 5).toDouble() },
+                IOType.d2(2, 2) { x, y -> (x * 2 + y + 1).toFloat() },
+                IOType.d2(2, 2) { x, y -> (x * 2 + y + 5).toFloat() },
             )
         val result = list.max()
         assertEquals(
@@ -63,7 +63,7 @@ class MaxExtTest {
         // [[1, 2],
         //  [3, 4],
         //  [5, 6]]
-        val a = IOType.d2(3, 2) { x, y -> (x * 2 + y + 1).toDouble() }
+        val a = IOType.d2(3, 2) { x, y -> (x * 2 + y + 1).toFloat() }
         val result = a.max(axis = 0)
         assertEquals(
             expected = IOType.d1(listOf(5.0, 6.0)),
@@ -76,7 +76,7 @@ class MaxExtTest {
         // [[1, 2],
         //  [3, 4],
         //  [5, 6]]
-        val a = IOType.d2(3, 2) { x, y -> (x * 2 + y + 1).toDouble() }
+        val a = IOType.d2(3, 2) { x, y -> (x * 2 + y + 1).toFloat() }
         val result = a.max(axis = 1)
         assertEquals(
             expected = IOType.d1(listOf(2.0, 4.0, 6.0)),
@@ -88,8 +88,8 @@ class MaxExtTest {
     fun `List_D2のmax_axis0=各D2の列方向最大値のリスト`() {
         val list =
             listOf(
-                IOType.d2(2, 2) { x, y -> (x * 2 + y + 1).toDouble() },
-                IOType.d2(2, 2) { x, y -> (x * 2 + y + 5).toDouble() },
+                IOType.d2(2, 2) { x, y -> (x * 2 + y + 1).toFloat() },
+                IOType.d2(2, 2) { x, y -> (x * 2 + y + 5).toFloat() },
             )
         val result = list.max(axis = 0)
         assertEquals(expected = 2, actual = result.size)
@@ -101,8 +101,8 @@ class MaxExtTest {
     fun `List_D2のmax_axis1=各D2の行方向最大値のリスト`() {
         val list =
             listOf(
-                IOType.d2(2, 2) { x, y -> (x * 2 + y + 1).toDouble() },
-                IOType.d2(2, 2) { x, y -> (x * 2 + y + 5).toDouble() },
+                IOType.d2(2, 2) { x, y -> (x * 2 + y + 1).toFloat() },
+                IOType.d2(2, 2) { x, y -> (x * 2 + y + 5).toFloat() },
             )
         val result = list.max(axis = 1)
         assertEquals(expected = 2, actual = result.size)
@@ -113,7 +113,7 @@ class MaxExtTest {
     @Test
     fun `D3のmax=全要素の最大値`() {
         // [[[1, 2], [3, 4]], [[5, 6], [7, 8]]]
-        val a = IOType.d3(2, 2, 2) { x, y, z -> (x * 4 + y * 2 + z + 1).toDouble() }
+        val a = IOType.d3(2, 2, 2) { x, y, z -> (x * 4 + y * 2 + z + 1).toFloat() }
         val result = a.max()
         assertEquals(
             expected = 8.0,
@@ -125,8 +125,8 @@ class MaxExtTest {
     fun `List_D3のmax=各D3の最大値のリスト`() {
         val list =
             listOf(
-                IOType.d3(2, 2, 2) { x, y, z -> (x * 4 + y * 2 + z + 1).toDouble() },
-                IOType.d3(2, 2, 2) { x, y, z -> (x * 4 + y * 2 + z + 10).toDouble() },
+                IOType.d3(2, 2, 2) { x, y, z -> (x * 4 + y * 2 + z + 1).toFloat() },
+                IOType.d3(2, 2, 2) { x, y, z -> (x * 4 + y * 2 + z + 10).toFloat() },
             )
         val result = list.max()
         assertEquals(
@@ -138,16 +138,16 @@ class MaxExtTest {
     @Test
     fun `D3のmax_axis0=各YZ平面の最大値`() {
         // [[[1, 2], [3, 4]], [[5, 6], [7, 8]]]
-        val a = IOType.d3(2, 2, 2) { x, y, z -> (x * 4 + y * 2 + z + 1).toDouble() }
+        val a = IOType.d3(2, 2, 2) { x, y, z -> (x * 4 + y * 2 + z + 1).toFloat() }
         val result = a.max(axis = 0)
-        val expected = IOType.d2(2, 2) { y, z -> (4 + y * 2 + z + 1).toDouble() }
+        val expected = IOType.d2(2, 2) { y, z -> (4 + y * 2 + z + 1).toFloat() }
         assertEquals(expected = expected, actual = result)
     }
 
     @Test
     fun `D3のmax_axis1=各XZ平面の最大値`() {
         // [[[1, 2], [3, 4]], [[5, 6], [7, 8]]]
-        val a = IOType.d3(2, 2, 2) { x, y, z -> (x * 4 + y * 2 + z + 1).toDouble() }
+        val a = IOType.d3(2, 2, 2) { x, y, z -> (x * 4 + y * 2 + z + 1).toFloat() }
         val result = a.max(axis = 1)
         // Expected: [[3, 4], [7, 8]]
         assertEquals(expected = 2, actual = result.shape[0])
@@ -161,7 +161,7 @@ class MaxExtTest {
     @Test
     fun `D3のmax_axis2=各XY平面の最大値`() {
         // [[[1, 2], [3, 4]], [[5, 6], [7, 8]]]
-        val a = IOType.d3(2, 2, 2) { x, y, z -> (x * 4 + y * 2 + z + 1).toDouble() }
+        val a = IOType.d3(2, 2, 2) { x, y, z -> (x * 4 + y * 2 + z + 1).toFloat() }
         val result = a.max(axis = 2)
         // Expected: [[2, 4], [6, 8]]
         assertEquals(expected = 2, actual = result.shape[0])
@@ -176,8 +176,8 @@ class MaxExtTest {
     fun `List_D3のmax_axis0=各D3のX軸方向最大値のリスト`() {
         val list =
             listOf(
-                IOType.d3(2, 2, 2) { x, y, z -> (x * 4 + y * 2 + z + 1).toDouble() },
-                IOType.d3(2, 2, 2) { x, y, z -> (x * 4 + y * 2 + z + 10).toDouble() },
+                IOType.d3(2, 2, 2) { x, y, z -> (x * 4 + y * 2 + z + 1).toFloat() },
+                IOType.d3(2, 2, 2) { x, y, z -> (x * 4 + y * 2 + z + 10).toFloat() },
             )
         val result = list.max(axis = 0)
         assertEquals(expected = 2, actual = result.size)
@@ -193,8 +193,8 @@ class MaxExtTest {
     fun `List_D3のmax_axis1=各D3のY軸方向最大値のリスト`() {
         val list =
             listOf(
-                IOType.d3(2, 2, 2) { x, y, z -> (x * 4 + y * 2 + z + 1).toDouble() },
-                IOType.d3(2, 2, 2) { x, y, z -> (x * 4 + y * 2 + z + 10).toDouble() },
+                IOType.d3(2, 2, 2) { x, y, z -> (x * 4 + y * 2 + z + 1).toFloat() },
+                IOType.d3(2, 2, 2) { x, y, z -> (x * 4 + y * 2 + z + 10).toFloat() },
             )
         val result = list.max(axis = 1)
         assertEquals(expected = 2, actual = result.size)
@@ -210,8 +210,8 @@ class MaxExtTest {
     fun `List_D3のmax_axis2=各D3のZ軸方向最大値のリスト`() {
         val list =
             listOf(
-                IOType.d3(2, 2, 2) { x, y, z -> (x * 4 + y * 2 + z + 1).toDouble() },
-                IOType.d3(2, 2, 2) { x, y, z -> (x * 4 + y * 2 + z + 10).toDouble() },
+                IOType.d3(2, 2, 2) { x, y, z -> (x * 4 + y * 2 + z + 1).toFloat() },
+                IOType.d3(2, 2, 2) { x, y, z -> (x * 4 + y * 2 + z + 10).toFloat() },
             )
         val result = list.max(axis = 2)
         assertEquals(expected = 2, actual = result.size)
